refactor: 블럭 상세 페이지 API DDD + hexagonal 로 리팩토링
