+++ src/main/java/com/joojoo/api/block/application/BlockServiceImpl.java
@@ -1,16 +1,12 @@
 package com.joojoo.api.block.application;
 
-import com.joojoo.api.block.application.detail.BlockDtoAssembler;
 import com.joojoo.api.block.application.validate.blockerTree.BlockTreeValidator;
 import com.joojoo.api.block.domain.model.entity.Block;
 import com.joojoo.api.block.domain.model.enums.DeleteMode;
 import com.joojoo.api.block.domain.repository.BlockRepository;
 import com.joojoo.api.block.presentation.dto.request.updateBlock.BlockUpdateDto;
-import com.joojoo.api.block.presentation.dto.response.detail.BlockDetailResponseDto;
 import com.joojoo.api.blockTag.domain.model.entity.BlockTag;
 import com.joojoo.api.blockTag.domain.repository.BlockTagRepository;
-import com.joojoo.api.blockTicker.domain.model.entity.BlockTicker;
-import com.joojoo.api.blockTicker.domain.repository.BlockTickerRepository;
 import com.joojoo.api.tag.domain.model.entity.Tag;
 import com.joojoo.api.util.metadata.MetadataService;
 import com.joojoo.global.exception.handleException.block.BlockNotFoundException;
@@ -31,23 +27,7 @@ public class BlockServiceImpl implements BlockService {
     private final BlockRepository blockRepository;
     private final BlockTreeValidator blockTreeValidator;
     private final MetadataService metadataService;
-
-    private final BlockDtoAssembler blockDtoAssembler;
     private final BlockTagRepository blockTagRepository;
-    private final BlockTickerRepository blockTickerRepository;
-
-    @Override
-    @Transactional(readOnly = true)
-    public BlockDetailResponseDto getBlockDetail(Long rootId) {
-        List<Block> blocks = blockRepository.findAllChildrenByRootId(rootId);
-        if (blocks.isEmpty()) throw new BlockNotFoundException();
-
-        List<Long> ids = blocks.stream().map(Block::getId).toList();
-        List<BlockTag> tags = blockTagRepository.findAllBlockTags(ids);
-        List<BlockTicker> tickers = blockTickerRepository.findAllBlockTickers(ids);
-
-        return blockDtoAssembler.assembleTree(rootId, blocks, tags, tickers);
-    }
 
     @Override
     @Transactional
